pyproject.toml
[project]
name = "tara-bot"
version = "0.1.0"
description = "Tara's personal AI bot — search flights + deals on Telegram"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
]
